Add HelloWorld test app with 10k dummy methods. Split benchmarks into separate classes, each for every layer of abstraction.
